#13 remove alphaess subpackage
