--- src/main/java/be/mathiasbosman/alphaessenergyidproxy/domain/ExportService.java
@@ -1,9 +1,0 @@
-package be.mathiasbosman.alphaessenergyidproxy.domain;
-
-/**
- * Export service interface.
- */
-public interface ExportService {
-
-  void exportStatisticsForPastWeek();
-}
