--- src/main/java/be/mathiasbosman/alphaessenergyidproxy/domain/PvStatistics.java
@@ -1,9 +1,0 @@
-package be.mathiasbosman.alphaessenergyidproxy.domain;
-
-/**
- * PvStatistics interface.
- */
-public interface PvStatistics {
-
-  double getPvTotal();
-}
