--- src/main/java/be/mathiasbosman/alphaessenergyidproxy/domain/alphaess/dto/LoginDto.java
@@ -1,13 +1,0 @@
-package be.mathiasbosman.alphaessenergyidproxy.domain.alphaess.dto;
-
-import be.mathiasbosman.alphaessenergyidproxy.config.AlphaessProperties.Credentials;
-
-/**
- * Dto to post to the authentication endpoint.
- */
-public record LoginDto(String username, String password) {
-
-  public static LoginDto fromCredentials(Credentials creds) {
-    return new LoginDto(creds.getUsername(), creds.getPassword());
-  }
-}
