+++ src/main/java/be/mathiasbosman/alphaessexport/domain/DataCollector.java
@@ -1,0 +1,12 @@
+package be.mathiasbosman.alphaessexport.domain;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+/**
+ * Data collector interface.
+ */
+public interface DataCollector {
+
+  Optional<PvStatistics> getPvStatistics(String identifier, LocalDateTime date);
+}
