+++ src/main/java/be/mathiasbosman/alphaessexport/domain/ExportService.java
@@ -1,0 +1,9 @@
+package be.mathiasbosman.alphaessexport.domain;
+
+/**
+ * Export service interface.
+ */
+public interface ExportService {
+
+  void exportStatisticsForPastWeek();
+}
