+++ src/main/java/be/mathiasbosman/alphaessexport/domain/PvStatistics.java
@@ -1,0 +1,9 @@
+package be.mathiasbosman.alphaessexport.domain;
+
+/**
+ * PvStatistics interface.
+ */
+public interface PvStatistics {
+
+  double getPvTotal();
+}
