+++ src/main/java/be/mathiasbosman/alphaessexport/domain/response/ResponseEntity.java
@@ -1,0 +1,20 @@
+package be.mathiasbosman.alphaessexport.domain.response;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Abstract response entity of the AlphaESS API.
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+public abstract class ResponseEntity {
+
+  private int code;
+  private String info;
+
+  @SuppressWarnings("unused") // forces a data field
+  abstract Object getData();
+}
