+++ src/main/java/be/mathiasbosman/alphaessexport/exporter/energyid/WebhookAdapter.java
@@ -1,0 +1,14 @@
+package be.mathiasbosman.alphaessexport.exporter.energyid;
+
+import java.net.URI;
+import org.springframework.http.HttpEntity;
+
+/**
+ * Interface for webhook adapter.
+ *
+ * @param <T> Type of the data to be posted.
+ */
+public interface WebhookAdapter<T> {
+
+  void postData(URI targetUrl, HttpEntity<T> httpEntity);
+}
