+++ src/main/java/be/mathiasbosman/alphaessexport/exporter/energyid/dto/MeterReadingDataDto.java
@@ -1,0 +1,12 @@
+package be.mathiasbosman.alphaessexport.exporter.energyid.dto;
+
+import java.time.LocalDateTime;
+
+/**
+ * Single data point of a reading. Currently, unused as the EnergyID webhook just takes an array of
+ * arrays.
+ */
+@SuppressWarnings("unused")
+public record MeterReadingDataDto(LocalDateTime date, double reading) {
+
+}
